requirements.txt
# IMA Lab 

# Core
pandas>=2.0.0
numpy>=1.24.0

# Data Loading
openpyxl>=3.1.0
gspread>=5.11.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# Database
# sqlite3 is built-in

# Visualization
plotly>=5.17.0
streamlit>=1.28.0

# Utils
toml>=0.10.2
